requirements.txt
transformers
Pillow
torch
streamlit
